requirements.txt
numpy>=1.26.4,<3.0
Pillow>=10.4.0
opencv-python>=4.10.0
scikit-learn>=1.5.1
torch>=2.4.1
torchvision>=0.19.1
ftfy>=6.3.1
regex>=2024.9.11
tqdm>=4.66.5
ultralytics>=8.3.0
transformers>=4.48.0
git+https://github.com/openai/CLIP.git

# Enhanced Detection Models
anthropic>=0.7.0  # Claude Vision API - Highest accuracy
groundingdino==0.1.0  # Open vocabulary detection
supervision>=0.19.0  # YOLO utilities
tabulate>=0.9.0  # Pretty table formatting

# Optional: Ollama support (requires local Ollama installation)
# Download from: https://ollama.ai
